requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.0
django-environ>=0.11.0
django-redis>=5.4.0
django-ratelimit>=4.1.0
Pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.9
python-dateutil>=2.8.2
boto3>=1.34.0
django-storages>=1.14.0
sentry-sdk>=1.40.0
reportlab>=4.0.0
pyotp>=2.9.0
qrcode>=7.4.2
stripe>=7.0.0

# WebSocket Support (Django Channels)
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Image Processing
Pillow>=10.0.0

# Geolocation & Timezone Support
pytz>=2024.1                  # Timezone database (IANA timezones)
geopy>=2.4.0                  # Geocoding (address -> lat/long, distance calculations)
django-timezone-field>=6.1.0  # Better timezone field for models
geoip2>=4.7.0                 # IP-based geolocation (MaxMind GeoIP2)
pycountry>=23.12.11           # ISO country/currency/language data
